Web App To Check which places are yet  to visit on map
Using Facebook connect user will logged in via FB and app will get all
the post which conatins location and that location will be mapped in
map.
Now User will able to see all places he has visited on map and when we
searches for any places on map he will get to know whether he has
visited that place early or he has not visited. If he has visited he
will get to know when he has visited with description and story what he
has posted on FB and the date when he has visited.
And with using mapmyindia POI api he can see all the POI Point Of
interest places on map which yet he has to visit in the area where he
lives and all which he has visited.

So here The people who are searching for any places on map and didn't
remember or know whether they have visited that place or not, they can
now instantly check and know all the visited and not visited places.